_python_archive/requirements.txt
flask>=2.3
anthropic>=0.25
openai>=1.12
python-dotenv>=1.0
